% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocument.R
\docType{data}
\name{data}
\alias{data}
\title{correlation coefficients between all pair companies}
\format{An object of class \code{data.frame} with 6 rows and 6 columns.}
\usage{
data
}
\description{
A dataset containing a  matrix of correlation coefficients between all pair companies. 6 row and 6 col.
}
\keyword{datasets}

