% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jdmbs.R
\name{normal_bs}
\alias{normal_bs}
\title{Normal A Monte Carlo Option Pricing Algorithm}
\usage{
normal_bs(companies, simulation.length = 180, monte_carlo = 1000,
  start_price, mu, sigma, K, color)
}
\arguments{
\item{companies}{is a number of simulate companies.}

\item{simulation.length}{is a duration of simulation.}

\item{monte_carlo}{is a iteration number of monte carlo.}

\item{start_price}{is a vector of initial price of company stock prices.}

\item{mu}{is a vector of parameters of geometric brown motions.}

\item{sigma}{is a voctor of parameters of geometric brown motions.}

\item{K}{is a vector of option execution prices.}

\item{color}{is a vector of colors in plot.}
}
\value{
premium of a list with (call_premium, put_premium)
}
\description{
Normal A Monte Carlo Option Pricing Algorithm
}
\examples{
normal_bs(1, simulation.length=180, monte_carlo=1000, 1000, 1, 0.6, 1200, "blue")
}

