% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jdmbs.R
\name{normal_bs}
\alias{normal_bs}
\title{Normal A Monte Carlo Option Pricing Algorithm}
\usage{
normal_bs(companies, simulation.length = 180, monte_carlo = 1000,
  start_price, mu, sigma, K, color)
}
\arguments{
\item{companies}{an integer of company number in order to simulate.}

\item{simulation.length}{an integer of a duration of simulation.}

\item{monte_carlo}{an integer of iterations of monte carlo.}

\item{start_price}{a vector of company's initial stock prices.}

\item{mu}{a vector of parameter of Geometric Brownian motion.}

\item{sigma}{a vector of parameter of Geometric Brownian motion.}

\item{K}{a vector of option execution prices.}

\item{color}{a vector of colors in plot.}
}
\value{
premium a list of (call_premium, put_premium)
}
\description{
Normal A Monte Carlo Option Pricing Algorithm
}
\examples{
normal_bs(1, simulation.length=50, monte_carlo=100,1000, 0.007, 0.03, 3000, "blue")
}

